Fix the ResourceResolver to FileName maPPING
Oops, the resource resolver is case senstive in Tomcat 8, but not 7.
That's the only explanation I can think of, as running the WAR in Tomcat 8
yields resource resolution failures, and running the WAR from the Tomcat
maven plugin (e.g. Tomcat 7), Just Works.

So, declaring the resource file in the servlet configuration as "Messages"
instead of "messages", so that I don't need to rename a file with case
only, on a case insensitive filesystem (which never ends well in git).