Migrate to SpringBoot runnable JAR
Switch from a WAR artifact to a JAR artifact that is a SpringBoot app
with embedded Jetty.  Change the way resource handlers are added and
Thymeleaf is configured, to find resources after having been moved to
worn with SpringBoot.nnChange templates to reference static assets
differently according to the new paths.