Switch from Spring MVC to Springboot (#3)
* Migrate to SpringBoot runnable JAR

Switch from a WAR artifact to a JAR artifact that is a SpringBoot app
with embedded Jetty.  Change the way resource handlers are added and
Thymeleaf is configured, to find resources after having been moved to
worn with SpringBoot.nnChange templates to reference static assets
differently according to the new paths.

* Adjust Heroku build commands to SpringBoot

Having moved from a WAR artifact to a JAR artifact, reflect the change
in the Heroku Procfiles which start the app after deployment.  Also,
remove the webapp runner gunk from the POM which was only a Heroku
requirement for WAR artifacts.